+++ pyimgur/exceptions.py
@@ -31,6 +31,10 @@ class UnexpectedImgurException(PyImgurError):
     or something else unexpected.
     """
 
+    def __init__(self, message, response=None):
+        super().__init__(message)
+        self.response = response
+
 
 class ImgurIsDownException(PyImgurError):
     """Imgur's API is not available."""